rename usersState container
