+++ netmanager/src/views/pages/SignUp/Login.js
@@ -8,7 +8,7 @@ import { clearErrors, loginUser } from "../../../redux/Join/actions";
 import classnames from "classnames";
 import { isEmpty, omit } from "underscore";
 import { isFormFullyFilled } from "./utils";
-import containerConnector from "../../stateConnectors/containerConnector";
+import usrsStateConnector from "views/stateConnectors/usersStateConnector";
 
 class Login extends Component {
   constructor(props) {
@@ -233,6 +233,6 @@ const mapStateToProps = (state) => ({
   auth: state.auth,
   errors: state.errors,
 });
-export default containerConnector(
+export default usrsStateConnector(
   connect(mapStateToProps, { clearErrors, loginUser })(Login)
 );
