+++ netmanager/src/views/pages/UserList/components/UsersTable/UsersTable.js
@@ -21,7 +21,7 @@ import {
 import { Alert, AlertTitle } from "@material-ui/lab";
 import { getInitials } from "helpers";
 import CustomMaterialTable from "views/components/Table/CustomMaterialTable";
-import containerConnector from "views/stateConnectors/containerConnector";
+import usersStateConnector from "views/stateConnectors/usersStateConnector";
 
 
 const roles = [
@@ -391,4 +391,4 @@ UsersTable.propTypes = {
   fetchUsers: PropTypes.func.isRequired,
 };
 
-export default containerConnector(UsersTable);
+export default usersStateConnector(UsersTable);
