+++ netmanager/src/views/pages/UserList/components/UsersToolbar/UsersToolbar.js
@@ -18,7 +18,7 @@ import { Alert, AlertTitle } from "@material-ui/lab";
 import { useMinimalSelectStyles } from "@mui-treasury/styles/select/minimal";
 import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
 import { useOrgData } from "redux/Join/selectors";
-import containerConnector from "views/stateConnectors/containerConnector";
+import usersStateConnector from "views/stateConnectors/usersStateConnector";
 
 
 const useStyles = makeStyles((theme) => ({
@@ -471,4 +471,4 @@ const mapStateToProps = (state) => ({
   auth: state.auth,
   errors: state.errors,
 });
-export default containerConnector(connect(mapStateToProps)(UsersToolbar));
+export default usersStateConnector(connect(mapStateToProps)(UsersToolbar));
